Refactoring
- Drop support for PHP < 7.2, and add support for PHP >= 8.
- Cleaned up configuration.
- Added PHPStorm config.
- Removed outdated tools and deps.
- Upgraded PHPUnit.
- Add meta and other project configs.
- Refactored tests.
- Add code quality tools.
- Add Github Actions config.
