+++ .env.example
@@ -1,0 +1,8 @@
+BASE_PATH=.
+BUILD_ROOT_PATH=/app
+PROJECT_NAME=dhii_transformer-interface
+
+PHP_BUILD_VERSION=7.2
+PHP_TEST_VERSION=7.2
+
+HOST_IP_ADDRESS=127.0.0.1
