+++ .gitignore
@@ -1,13 +1,9 @@
 /vendor/
-/nbproject/private/
-/test/coverage/
-/test/log/
 /build/
+/.idea/workspace.xml
 /.idea/codeStyles/
 /.idea/inspectionProfiles/
-/.idea/dictionaries/
-/.idea/workspace.xml
-/.idea/tasks.xml
-/.idea/deployment.xml
-/.idea/vagrant.xml
 /.idea/misc.xml
+/.env
+
+/.phpunit.result.cache
