+++ .idea/php-test-framework.xml
@@ -5,7 +5,7 @@
       <tool tool_name="PHPUnit">
         <cache>
           <versions>
-            <info id="Local" version="4.8.36" />
+            <info id="interpreter-ba314a65-6074-40c9-8112-007ea3db0a5f" version="8.5.21" />
           </versions>
         </cache>
       </tool>
