+++ composer.json
@@ -1,45 +1,32 @@
 {
-    "name": "dhii/transformer-interface",
-    "description": "Interfaces for data transformation",
-    "type": "library",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Dhii Team",
-            "email": "development@dhii.co"
-        }
-    ],
-    "minimum-stability": "dev",
-    "prefer-stable": true,
-    "require": {
-        "php": "^5.3 | ^7.0",
-        "dhii/exception-interface": "^0.1 | ^0.2-alpha1",
-        "dhii/factory-interface": "^0.1-alpha1"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^4.8",
-        "ptrofimov/xpmock": "^1.1",
-        "dhii/php-cs-fixer-config": "dev-php-5.3",
-        "codeclimate/php-test-reporter": "<=0.3.2"
-    },
-    "autoload": {
-        "psr-4": {
-            "Dhii\\Transformer\\": "src"
-        }
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "Dhii\\Transformer\\UnitTest\\": "test/unit",
-            "Dhii\\Transformer\\FuncTest\\": "test/functional"
-        }
-    },
-    "scripts": {
-        "test": "phpunit",
-        "csfix": "php-cs-fixer fix -vvv"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-develop": "0.1.x-dev"
-        }
+  "name": "dhii/transformer-interface",
+  "description": "Interfaces for data transformation",
+  "type": "library",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Dhii Team",
+      "email": "development@dhii.co"
     }
+  ],
+  "minimum-stability": "dev",
+  "prefer-stable": true,
+  "require": {
+    "php": "^7.2|^8.0"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^7.0|^8.0|^9.0",
+    "vimeo/psalm": "^4.0",
+    "slevomat/coding-standard": "^6.0"
+  },
+  "autoload": {
+    "psr-4": {
+      "Dhii\\Transformer\\": "src"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "Dhii\\Transformer\\UnitTest\\": "tests/unit"
+    }
+  }
 }
