--- phpmd.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0"?>
-
-<ruleset name="Dhii Standard"
-         xmlns="http://pmd.sf.net/ruleset/1.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://pmd.sf.net/ruleset/1.0.0
-                             http://pmd.sf.net/ruleset_xml_schema.xsd"
-         xsi:noNamespaceSchemaLocation="http://pmd.sf.net/ruleset_xml_schema.xsd">
-    <description>A standard Dhii ruleset</description>
-
-    <!-- Standard Rules of CodeClimate -->
-    <rule ref="rulesets/cleancode.xml" />
-    <rule ref="rulesets/codesize.xml" />
-    <rule ref="rulesets/design.xml" />
-    <rule ref="rulesets/naming.xml" />
-    <rule ref="rulesets/unusedcode.xml" />
-
-
-    <rule ref="rulesets/controversial.xml/CamelCaseMethodName">
-        <properties>
-            <property name="allow-underscore" value="true" />
-        </properties>
-    </rule>
-</ruleset>
