--- phpunit.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<phpunit
-    colors="true"
-    bootstrap="test/bootstrap.php"
->
-    <php>
-        <ini name="display_errors" value="1" />
-        <ini name="display_startup_errors" value="1" />
-    </php>
-    <testsuites>
-        <testsuite name="unit">
-            <directory>./test/unit/</directory>
-        </testsuite>
-        <testsuite name="functional">
-            <directory>./test/functional/</directory>
-        </testsuite>
-    </testsuites>
-    <filter>
-        <whitelist processUncoveredFilesFromWhitelist="true">
-            <directory suffix=".php">src</directory>
-        </whitelist>
-    </filter>
-    <logging>
-        <log type="coverage-html" target="./test/coverage/html" lowUpperBound="35"
-             highLowerBound="80"/>
-        <log type="coverage-clover" target="./test/coverage/clover.xml"/>
-        <log type="coverage-php" target="./test/coverage/serialized"/>
-        <log type="coverage-text" target="php://stdout" showUncoveredFiles="false"/>
-        <log type="json" target="./test/log/logfile.json"/>
-        <log type="tap" target="./test/log/logfile.tap"/>
-        <log type="junit" target="./test/log/logfile.xml" logIncompleteSkipped="false"/>
-        <log type="testdox-html" target="./test/log/testdox.html"/>
-        <log type="testdox-text" target="./test/log/testdox.txt"/>
-    </logging>
-</phpunit>
