+++ phpunit.xml.dist
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<phpunit
+    colors="true"
+    bootstrap="tests/bootstrap.php"
+>
+    <php>
+        <ini name="display_errors" value="1" />
+        <ini name="display_startup_errors" value="1" />
+    </php>
+    <testsuites>
+        <testsuite name="functional">
+            <directory>tests/unit</directory>
+        </testsuite>
+    </testsuites>
+</phpunit>
