+++ src/TransformerAwareInterface.php
@@ -1,20 +1,17 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Dhii\Transformer;
 
 /**
  * Something that exposes a transformer.
- *
- * @since [*next-version*]
  */
 interface TransformerAwareInterface
 {
     /**
      * Retrieves the transformer associated with this instance.
      *
-     * @since [*next-version*]
-     *
-     * @return TransformerInterface|null
      */
-    public function getTransformer();
+    public function getTransformer(): ?TransformerInterface;
 }
