--- test/bootstrap.php
@@ -1,5 +1,0 @@
-<?php
-
-error_reporting(E_ALL | E_STRICT);
-
-require_once dirname(__FILE__).'/../vendor/autoload.php';
