+++ tests/bootstrap.php
@@ -1,0 +1,5 @@
+<?php
+
+error_reporting(E_ALL | E_STRICT);
+
+require_once dirname(__FILE__).'/../vendor/autoload.php';
